第一次整个workflow 初步搞定
这个版本可以直接使用ziyan_make命令 生成相应工作目录，然后运行python manage.py 进行redis入队。需要driftluo 添加python manage.py run 功能
