--- manage.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-from ziyan.ziyan_main import start
-import os
-print(os.getcwd())
-start()
-while True:
-    pass
