+++ ziyan/script/manage.py
@@ -1,1 +1,6 @@
 # -*- coding: utf-8 -*-
+import ziyan
+
+ziyan.start()
+while True:
+    pass
